rename sidebarnav to index
